[Angus] Show FMS nav correctly on all pages
I've put the FMS menu HTML in its own template so it can be included in two
separate places in the DOM: the usual place in the footer (for use on mobile),
and just before the main page content. Doing it this way seems simpler than
abusing CSS to put the nav beneath the Angus breadcrumbs.